+++ wayland/tkWaylandButton.c
@@ -184,7 +184,7 @@ DrawButtonBitmap(TkButton *butPtr,
     if (!bitmap) {
         /* No bitmap: draw fallback rectangle. */
         nvgBeginPath(dc->vg);
-        nvgRect(dc->vg, x, y, width, height);
+        nvgRect(dc->vg, (float) x, (float) y, (float) width, (float) height);
         nvgFillColor(dc->vg, nvgRGBA(192, 192, 192, 255));
         nvgFill(dc->vg);
         return;
@@ -289,7 +289,8 @@ DrawButtonBitmap(TkButton *butPtr,
     if (imageId > 0) {
         NVGpaint paint = nvgImagePattern(dc->vg, x, y, bm_width, bm_height, 0, imageId, 1);
         nvgBeginPath(dc->vg);
-        nvgRect(dc->vg, x, y, bm_width, bm_height);
+        nvgRect(dc->vg, (float) x, (float) y,
+		(float) bm_width, (float) bm_height);
         nvgFillPaint(dc->vg, paint);
         nvgFill(dc->vg);
         nvgDeleteImage(dc->vg, imageId);
@@ -309,7 +310,7 @@ cleanup:
 
 fallback_rect:
     nvgBeginPath(dc->vg);
-    nvgRect(dc->vg, x, y, width, height);
+    nvgRect(dc->vg, (float) x, (float) y, (float) width, (float) height);
     nvgFillColor(dc->vg, nvgRGBA(192, 192, 192, 255));
     nvgFill(dc->vg);
     return;
@@ -991,7 +992,7 @@ TkpDrawCheckIndicator(
     
     /* Draw background. */
     nvgBeginPath(dc->vg);
-    nvgRect(dc->vg, x, y, size, size);
+    nvgRect(dc->vg, (float) x, (float) y, (float) size, (float) size);
     
     if (disabled && disColor) {
         nvgFillColor(dc->vg, TkGlfwXColorToNVG(disColor));
